fix(ci): coverage-shards — relocate validate-only recovery head helper
The task-tag feature pushed execution_flow.py to 1502 lines, tripping the
1500-line maintainability guardrail. Move the pure
_validate_only_recovery_target_head_sha helper to the sibling
recovery_payloads module (its natural home alongside the other recovery
payload helpers) and re-import it, keeping the public import path and
behavior unchanged. execution_flow.py is now 1483 lines.
